requirements.txt
pywin32>=311
pycryptodomex>=3.23.0
psutil>=5.9.0
